requirements.txt
streamlit
pytesseract
pdf2image
langdetect
python-docx
reportlab
Pillow
